requirements.txt
django=1.8.5
djangorestframework=3.6.4
-e git+https://github.com/jhuapl-boss/django-oidc.git#egg=django-oidc
-e git+https://github.com/jhuapl-boss/drf-oidc-auth.git#egg=drf-oidc-auth
-e git+https://github.com/jhuapl-boss/boss-oidc.git#egg=boss-oidc